Responsable: Nahuel Alberto Cachambi user: CachaSan10 Descripción: Se creo la implementacion de service de consejo para manipular los datos de consejo en la base de datos: ConsejoServiceMysqlImp.Se implemento el service ConsejoServiceMysql en el controller de consejo. Se creo un archivo import.sql para que precargue datos a la base de datos.Solamente contiene el archivo un ejemplar de consejo. Close #138 Close #146

+++ Tp2_Grupo30/src/main/java/ar/edu/unju/fi/service/IConsejoService.java
@@ -15,11 +15,11 @@ public interface IConsejoService {
 	
 	public void addConsejo(Consejo consejo,MultipartFile imagen) throws IOException;
 	
-	public Consejo getConsejoEncontrado(int id);
+	public Consejo getConsejoEncontrado(Long id);
 	
 	
 	public void updateConsejo(Consejo consejoModificado,MultipartFile imagen) throws IOException;
 	
-	public void deleteConsejo(int id);
+	public void deleteConsejo(Long id);
 	
 }
